# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BUCT Media HUB",
    "description": "北京化工大学媒体资源管理系统后端接口。\n\n认证方式：大多数受保护接口使用 `Authorization: Bearer <token>`。\n前端常用登录接口为 `POST /api/v1/auth/login`，OpenAPI 工具链兼容登录接口为 `POST /api/v1/auth/token`。",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/stats/view/{photo_id}": {
      "post": {
        "tags": [
          "Statistics"
        ],
        "summary": "Increment View",
        "description": "Increment view count for a photo",
        "operationId": "increment_view_api_v1_stats_view__photo_id__post",
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats/dashboard": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Get Dashboard Stats",
        "description": "Get dashboard statistics (reviewer only)",
        "operationId": "get_dashboard_stats_api_v1_stats_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        },
        "description": "使用 `POST /api/v1/auth/token` 获取 Bearer Token，或通过 `POST /api/v1/auth/login` 获取前端登录响应。"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地开发环境"
    },
    {
      "url": "/",
      "description": "相对路径，适合反向代理或 ApiFox 自定义环境"
    }
  ],
...
}